refactor: moved api wrapper code into it's own internal dir
